requirements.txt
datasets
pandas
flask
flask-cors
groq
python-dotenv
gunicorn
redis
psycopg2-binary
sqlalchemy
marshmallow
pytest
pytest-cov
black
flake8
psutil
matplotlib
plotly
dash
schedule
numpy
scikit-learn
